+++ ICD.Connect.Partitioning.Commercial/CallRatings/CallRatingTelemetryNames.cs
@@ -2,7 +2,10 @@
 {
     public static class CallRatingTelemetryNames
     {
-	    public const string CALL_RATING_AVERAGE = "CallRatingAverage";
+	    public const string CALL_RATING_AVERAGE_STRING = "CallRatingAverageString";
 	    public const string CALL_RATING_AVERAGE_CHANGED = "CallRatingAverageChanged";
+
+	    public const string RECENT_CALL_RATINGS_CHANGED = "OnRecentCallRatingsChanged";
+	    public const string RECENT_CALL_RATINGS = "RecentCallRatings";
     }
 }