+++ ICD.Connect.Partitioning.Commercial/Properties/AssemblyInfo.cs
@@ -4,4 +4,4 @@
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Partitioning.Commercial")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2021")]
-[assembly: AssemblyVersion("15.4.0.0")]
+[assembly: AssemblyVersion("15.4.1.0")]