+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added CalendarManager Callbacks to AbstractCommercialRoom.
  - Added RoomType property to ICommercialRoom and AbstractCommercialRoom
  - Added OperationalHours to ICommercialRoom and AbstractCommercialRoom
+ - Added CallRatingManager to ICommercialRoom and AbstractCommercailRoom
 
 ## [15.3.0] - 2020-09-28
 ### Added